+++ src/i18n/types/resources.ts
@@ -11,7 +11,8 @@ const {
     actionBar,
     pnrSearch,
     filterSort,
-    auth
+    landingPage,
+    auth,
 } = data;
 const resources = {
     ns1,
@@ -24,6 +25,7 @@ const resources = {
     pnrSearch,
     filterSort,
     tripDetails,
+    landingPage,
     passengerDetails,
 } as const;