+++ src/pages/TripsListingPage/TripsListingPage.tsx
@@ -1,9 +1,12 @@
 import { setTripListingData } from '../../app/features/tripListingSlice';
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { TripCardAccordion } from '../../components';
+import ActionBarDrawer from '../../components/actionBar/actionBarDrawer/ActionBarDrawer';
+import ActionBarTab from '../../components/actionBar/actionBarTab/ActionBarTab';
 import LoadMore from '../../components/loadMore/LoadMore';
 import { ISeatType, IBusType, ITrip } from '../../types';
 import { TripsListingPageWrapper } from './TripsListingPage.styled';
+import useMediaQuery from '@mui/material/useMediaQuery';
 
 // const dummyData = [
 //     {
@@ -150,14 +153,17 @@ const dummyData: ITrip[] = [
         totalSeats: 30,
     },
 ];
+
 export const TripsListingPage = () => {
     const state = useAppSelector((state) => state.tripListing);
     const dispatch = useAppDispatch();
+    const matches = useMediaQuery('(min-width:600px)');
+
     dispatch(setTripListingData(dummyData));
 
     return (
         <TripsListingPageWrapper>
-            <h1>TripsListingPage</h1>
+            {matches ? <ActionBarTab showFilterSort /> : <ActionBarDrawer />}
             {state.map((indData) => (
                 <TripCardAccordion key={indData.id} data={indData} />
             ))}