+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 -   [86cugqgpt] Implemented filters and sort for action bar
 -   [86cueubmz] Implemented trip booking page and booking details form UI
 -   [86cukqvct] Implemented API layer
+-   [86cugexhf] Landing page illustration, action bar incorporated and routing to trip listing page.